Improved Path + UNTESTED EDIT
Slightly improved initial turn and angle (decreased angle, increased
encoder distance, comes in more at the angle it used to now.) I tried
increasing difference btw motor powers in movetoWall method BUT that
increased the amount of right side turning. I remembered that the RIGHT
SIDE motors are actually hardware mapped to the teleOp front, so left
motors for auto. I flipped the motor powers in movetoWall but was unable
to test due to driver practice.